Optimize sameOrSubdomain where is safe to
