+++ app/src/main/java/com/duckduckgo/app/browser/pageloadpixel/PageLoadedHandler.kt
@@ -16,6 +16,7 @@
 
 package com.duckduckgo.app.browser.pageloadpixel
 
+import android.util.Log
 import com.duckduckgo.app.di.AppCoroutineScope
 import com.duckduckgo.browser.api.WebViewVersionProvider
 import com.duckduckgo.common.utils.DispatcherProvider
@@ -55,6 +56,7 @@ class RealPageLoadedHandler @Inject constructor(
     override operator fun invoke(url: String, start: Long, end: Long) {
         appCoroutineScope.launch(dispatcherProvider.io()) {
             if (sites.any { UriString.sameOrSubdomain(url, it) }) {
+                Log.v("PageLoadedLog", (end-start).toString())
                 pageLoadedPixelDao.add(
                     PageLoadedPixelEntity(
                         elapsedTime = end - start,
